chore(refactor): separate elements into separate components (#119) (#119)
* chore: update name

* chore(refactor): seperate desktop and mobile views

* chore(refactor): seperate desktop and mobile date picker view (wip)

* chore(refactor): fix mobile date picker bug and refactor code

* chore(refactor): lift state for mobile and desktop filter views

* chore(refactor): separate elements into separate components

* chore(refactor): file structure

* chore(refactor): change schema title to name
