+++ src/common/components/ProjectsFilter/svgs.ts
@@ -1,0 +1,7 @@
+import Back from 'src/assets/icons/Back'
+import Down from '../../../assets/icons/Down'
+import Reset from 'src/assets/icons/Reset'
+import Filter from '../../../assets/icons/Filter'
+import CalendarSort from 'src/assets/icons/CalendarSort'
+
+export { Back, Down, Reset, Filter, CalendarSort }
